Handle deeplinks in sign-client
