+++ packages/sign-client/package.json
@@ -16,6 +16,7 @@
     "wallet",
     "walletconnect"
   ],
+  "sideEffects": false,
   "scripts": {
     "clean": "rm -rf dist",
     "build:pre": "npm run clean",
