+++ packages/utils/src/misc.ts
@@ -6,6 +6,7 @@ import {
   EngineTypes,
   RelayerTypes,
 } from "@walletconnect/types";
+import type { IKeyValueStorage } from "@walletconnect/keyvaluestorage";
 import { getDocument, getLocation, getNavigator } from "@walletconnect/window-getters";
 import { getWindowMetadata } from "@walletconnect/window-metadata";
 import { ErrorResponse } from "@walletconnect/jsonrpc-utils";
@@ -320,4 +321,44 @@ export function engineEvent(event: EngineTypes.Event, id?: number | string | und
 
 export function mergeArrays<T>(a: T[] = [], b: T[] = []): T[] {
   return [...new Set([...a, ...b])];
+}
+
+export async function handleDeeplinkRedirect({
+  id,
+  topic,
+  store,
+}: {
+  id: number;
+  topic: string;
+  store: IKeyValueStorage;
+}) {
+  try {
+    const item = await store.getItem("WALLETCONNECT_DEEPLINK_CHOICE");
+
+    if (!item) return;
+
+    const json = typeof item === "string" ? JSON.parse(item) : item;
+    let deeplink = json?.href;
+
+    if (typeof deeplink !== "string") return;
+
+    if (deeplink.endsWith("/")) deeplink = deeplink.slice(0, -1);
+
+    const link = `${deeplink}/wc?requestId=${id}&sessionTopic=${topic}`;
+
+    const env = getEnvironment();
+
+    if (env === ENV_MAP.browser) {
+      window.open(link, "_self", "noreferrer noopener");
+    } else if (env === ENV_MAP.reactNative) {
+      // global.Linking is set by react-native-compat
+      if (typeof (global as any)?.Linking !== "undefined") {
+        await (global as any).Linking.openURL(link);
+      }
+    }
+  } catch (err) {
+    // Silent error, just log in console
+    // eslint-disable-next-line no-console
+    console.error(err);
+  }
 }
