+++ providers/universal-provider/src/providers/cosmos.ts
@@ -11,7 +11,7 @@ import {
   SessionNamespace,
   SubProviderOpts,
 } from "../types";
-import { getGlobal, getRpcUrl, handleDeepLinks } from "../utils";
+import { getGlobal, getRpcUrl } from "../utils";
 
 class CosmosProvider implements IProvider {
   public name = "cosmos";
@@ -50,7 +50,6 @@ class CosmosProvider implements IProvider {
 
   public request<T = unknown>(args: RequestParams): Promise<T> {
     if (this.namespace.methods.includes(args.request.method)) {
-      handleDeepLinks(this.client, args);
       return this.client.request(args as EngineTypes.RequestParams);
     }
     return this.getHttpProvider().request(args.request);
