+++ providers/universal-provider/src/providers/solana.ts
@@ -11,7 +11,7 @@ import {
   SessionNamespace,
   SubProviderOpts,
 } from "../types";
-import { getGlobal, getRpcUrl, handleDeepLinks } from "../utils";
+import { getGlobal, getRpcUrl } from "../utils";
 
 class SolanaProvider implements IProvider {
   public name = "solana";
@@ -39,7 +39,6 @@ class SolanaProvider implements IProvider {
 
   public request<T = unknown>(args: RequestParams): Promise<T> {
     if (this.namespace.methods.includes(args.request.method)) {
-      handleDeepLinks(this.client, args);
       return this.client.request(args as EngineTypes.RequestParams);
     }
     return this.getHttpProvider().request(args.request);
